+++ src/runtime/nw4r/src/nw4r/ef/emform/ef_cylinder.cpp
@@ -27,8 +27,8 @@ void EmitterFormCylinder::EmissionSub(
             angle = (pParams[3] - pParams[2]) * pEmitter->mRandom.RandFloat();
         }
 
-        f32 sa = std::sinf(angleOffset + angle);
-        f32 ca = std::cosf(angleOffset + angle);
+        f32 sa = std::sin(angleOffset + angle);
+        f32 ca = std::cos(angleOffset + angle);
 
         math::VEC3 fromYAxis(sa, 0.0f, -ca);
 
@@ -111,7 +111,7 @@ void EmitterFormCylinder::Emission(Emitter* pEmitter, ParticleManager* pManager,
     }
 
     if (flags & EmitterDesc::EMIT_FLAG_17) {
-        f32 f = std::fmodf(pParams[3] - pParams[2], NW4R_MATH_PI * 2);
+        f32 f = std::fmod(pParams[3] - pParams[2], static_cast<f32>(NW4R_MATH_PI * 2));
 
         if (f < NW4R_EF_EMIT_ANGLE_MIN || f > NW4R_EF_EMIT_ANGLE_MAX) {
             dangle =
