+++ src/runtime/nw4r/src/nw4r/ef/emform/ef_disc.cpp
@@ -38,7 +38,7 @@ void EmitterFormDisc::Emission(Emitter* pEmitter, ParticleManager* pManager,
     }
 
     if (flags & EmitterDesc::EMIT_FLAG_17) {
-        f32 f = std::fmodf(pParams[3] - pParams[2], NW4R_MATH_PI * 2);
+        f32 f = std::fmod(pParams[3] - pParams[2], static_cast<f32>(NW4R_MATH_PI * 2));
 
         if (f < NW4R_EF_EMIT_ANGLE_MIN || f > NW4R_EF_EMIT_ANGLE_MAX) {
             dangle =
@@ -59,7 +59,7 @@ void EmitterFormDisc::Emission(Emitter* pEmitter, ParticleManager* pManager,
 
         if (flags & EmitterDesc::EMIT_FLAG_24) {
             dist = dist + inner * inner * (1.0f - dist);
-            dist = std::sqrtf(dist);
+            dist = std::sqrt(dist);
         } else {
             dist = dist + inner * (1.0f - dist);
         }
@@ -68,8 +68,8 @@ void EmitterFormDisc::Emission(Emitter* pEmitter, ParticleManager* pManager,
             angle = (pParams[3] - pParams[2]) * pEmitter->mRandom.RandFloat();
         }
 
-        f32 sa = std::sinf(angleOffset + angle);
-        f32 ca = std::cosf(angleOffset + angle);
+        f32 sa = std::sin(angleOffset + angle);
+        f32 ca = std::cos(angleOffset + angle);
 
         math::VEC3 fromYAxis(sa, 0.0f, -ca);
 
@@ -87,9 +87,9 @@ void EmitterFormDisc::Emission(Emitter* pEmitter, ParticleManager* pManager,
         } else {
             f32 cone = dist * pEmitter->mParameter.mVelDiffusionEmitterNormal;
 
-            normal.x = std::sinf(cone) * sa;
-            normal.y = std::cosf(cone);
-            normal.z = std::sinf(cone) * -ca;
+            normal.x = std::sin(cone) * sa;
+            normal.y = std::cos(cone);
+            normal.z = std::sin(cone) * -ca;
         }
 
         CalcVelocity(&vel, pEmitter, pos, normal, fromOrigin, fromYAxis);
